requirements.txt
Flask==3.0.0
google-cloud-datastore==2.20.1
google-cloud-storage==2.19.0
python-jose
six
requests
authlib
